requirements.txt
apache-airflow==2.10.3
apache-airflow-providers-postgres==5.13.1
apache-airflow-providers-smtp==1.8.0
PyYAML==6.0.2
python-dotenv==1.1.1
SQLAlchemy==1.4.54
psycopg2-binary==2.9.10
PyMySQL==1.1.1
pytest==8.4.1
